+++ src/types/editorjs/Core.ts
@@ -4,7 +4,6 @@ import { InstallableTool } from './installableTool.js';
  * Class for editor.js core
  */
 export class Core extends InstallableTool {
-
   /**
    * Initiate editor.js core
    *
